requirements.txt
python-dotenv
prometheus_client
# pandas